reorganised

+++ init_my_system.sh
@@ -13,20 +13,31 @@ echo "current machine is ${machine}"
 
 the_dir=$(dirname $(pwd)/$0)
 
-# tmux
-curl -L https://raw.githubusercontent.com/gpakosz/.tmux/master/.tmux.conf -o ~/.tmux.conf
-ln -s ${the_dir}/tmux/tmux/ ~/.tmux
-ln -s ${the_dir}/tmux/tmux.conf ~/.tmux.conf.local
+# zsh
+if [ ! -f ~/.zshrc ];then
+	ln -s ${the_dir}/zsh/zshrc ~/.zshrc
+	ln -s ${the_dir}/zsh/zsh_env ~/.zsh_env
+else
+	echo "zshrc exists already"
+fi
 
-# oh-my-zsh
-git clone https://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
-ln -s ${the_dir}/zsh/zshrc ~/.zshrc
-ln -s ${the_dir}/zsh/zsh_env ~/.zsh_env
 if [ -f ~/.zsh_local ];then
 	echo "~/.zsh_local already exist. please write any local configurations into this file"
 else
 	echo "# please write local configuration into .zsh_local file" | tee ~/.zsh_local
+	echo "export PATH=$HOME/.local/bin:$PATH" >> ~/.zsh_local
 fi
+
+# tmux
+if [ ! -f ~/.tmux.conf ];then
+	curl -L https://raw.githubusercontent.com/gpakosz/.tmux/master/.tmux.conf -o ~/.tmux.conf
+	ln -s ${the_dir}/tmux/tmux.conf ~/.tmux.conf.local
+	mkdir -p ~/.tmux/plugins
+	git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+else
+	echo "tmux conf exists"
+fi
+
 
 # vim & neovim
 mkdir -p ~/.vimtmp
@@ -38,15 +49,16 @@ ln -s ${the_dir}/vim/vimrc ~/.vimrc
 ln -s ${the_dir}/vim/nvim ~/.config/nvim
 mkdir -p ~/.local/bin/
 
-if [ "$machine" = "Linux" ]; then
-	curl -LO https://github.com/neovim/neovim/releases/download/nightly/nvim.appimage
-	chmod u+x nvim.appimage
-	mv nvim.appimage $HOME/.local/bin/vim
+
+if [ -f ~/.local/bin/vim ];then
+	echo "~/.local/bin/vim exists" 
 else
-	echo "appimage is only working for Linux system, please find https://github.com/neovim/neovim/wiki/Installing-Neovim for Neovim installation" 
+	if [ "$machine" = "Linux" ]; then
+		curl -LO https://github.com/neovim/neovim/releases/download/nightly/nvim.appimage
+		chmod u+x nvim.appimage
+		mv nvim.appimage $HOME/.local/bin/vim
+	else
+		echo "appimage is only working for Linux system, please find https://github.com/neovim/neovim/wiki/Installing-Neovim for Neovim installation" 
+	fi
 fi
 
-curl -LO https://github.com/neovim/neovim/releases/download/nightly/nvim.appimage
-chmod u+x nvim.appimage
-mv nvim.appimage $HOME/.local/bin/vim
-echo "export PATH=$HOME/.local/bin:$PATH" >> ~/.zsh_local
